requirements.txt
opencage
mysql-connector-python
Faker
pandas